+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 ## 1.0.0 - 2022-12-16
 
 * First Release
+
+
+## 1.0.1 - 2023-01-19
+
+* Moved entities to the respective directory at Models/Entities
+* Added setter/getters to ClientConfig and OAuth2Config classes
+* Added OTPOptionEntity model
+* Added the missed "hash" property to UserEntity model